requirements.txt
archive==0.3
dm_clrs==1.0.0
loguru==0.7.2
matplotlib==3.7.1
networkx==3.1
numpy==1.24.3
torch==2.1.0
torch_geometric==2.5.1
tqdm==4.65.0